@@ -5,14 +5,14 @@
   "dependencies": {
     "cookie": "^0.4.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.1",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^13.0.1",
     "@testing-library/user-event": "^14.0.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
