@@ -1,11 +1,11 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
 import * as serviceWorker from './serviceWorker'
 import cookie from 'cookie'
 import { mergeFeatures } from './FeatureToggle'
 import defaultFeatures from './features'
+import { createRoot } from 'react-dom/client'
 
 const cookies = cookie.parse(document.cookie)
 const cookieFeatures = (cookies.features || '')
@@ -14,11 +14,11 @@ const cookieFeatures = (cookies.features || '')
 
 const features = mergeFeatures(defaultFeatures, cookieFeatures)
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'))
+root.render(
   <React.StrictMode>
     <App features={features} />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
 
 serviceWorker.unregister()
